grid cleanup, new number unit, split up stats
