+++ src/components/pages/Account/Stats.module.css
@@ -1,0 +1,6 @@
+.stats {
+  display: grid;
+  gap: var(--spacer);
+  grid-template-columns: repeat(auto-fit, minmax(10rem, 1fr));
+  margin-top: var(--spacer);
+}
